+++ src/Cookie/SetCookie.php
@@ -434,11 +434,12 @@ class SetCookie
         // Remove the leading '.' as per spec in RFC 6265.
         // https://datatracker.ietf.org/doc/html/rfc6265#section-5.2.3
         $cookieDomain = \ltrim(\strtolower($cookieDomain), '.');
+        if ('' === $cookieDomain) {
+            return false;
+        }
 
         $domain = \strtolower($domain);
-
-        // Domain not set or exact match.
-        if ('' === $cookieDomain || $domain === $cookieDomain) {
+        if ($domain === $cookieDomain) {
             return true;
         }
 
@@ -491,7 +492,7 @@ class SetCookie
         // Domains must not be empty, but may be omitted. "0" is not a valid
         // internet domain, but may be used as server name in a private network.
         $domain = $this->getDomain();
-        if ($domain === '') {
+        if ($domain === '' || (null !== $domain && '' === \ltrim(\trim($domain), '.'))) {
             return 'The cookie domain must not be empty';
         }